requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bitsandbytes==0.46.1
cachetools==6.2.4
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
dnspython==2.7.0
easyocr==1.7.2
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
git-filter-repo==2.47.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inference-sdk==0.51.2
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.3
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
postgrest==2.27.0
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.3
pyroaring==1.0.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.27.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
sentry-sdk==2.33.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.2
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
supervision==0.25.1
sympy==1.13.3
tenacity==9.1.2
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchvision==0.22.1+cu126
tqdm==4.67.1
transformers==4.53.1
trl==0.19.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.22.0
